# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Descripcion del API para el submodulo de rutas",
    "version": "1.0.0",
    "title": "Ms-rutas"
  },
  "host": "100.26.181.55:80",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/relationdateroute/createRelationDayRoute": {
      "post": {
        "tags": [
          "Relacion Dia-Ruta"
        ],
        "summary": "Agregar una relacion entre un dia y una ruta",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ids de los objetos dia y ruta",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "integer"
                },
                "idDay": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idRelationDateRoute": {
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/relationdateroute/searchRelationDayRoute": {
      "post": {
        "tags": [
          "Relacion Dia-Ruta"
        ],
        "summary": "Busca una relacion entre un dia y una ruta",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ids de los objetos dia y ruta",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "integer"
                },
                "idDay": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idRelationDateRoute": {
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/relationdateroute/deleteRelationDayRoute": {
      "delete": {
        "tags": [
          "Relacion Dia-Ruta"
        ],
        "summary": "Elimina una relacion entre un dia y una ruta",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ids de los objetos dia y ruta",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "integer"
                },
                "idDay": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/relationcardate/createRelationCarDay": {
      "post": {
        "tags": [
          "Relacion Auto-Dia"
        ],
        "summary": "Crea una relacion entre un auto y un dia",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ids de los objetos dia y auto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idCar": {
                  "type": "integer"
                },
                "idDay": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idRelationCarDate": {
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/relationcardate/getRelationByDate": {
      "post": {
        "tags": [
          "Relacion Auto-Dia"
        ],
        "summary": "Obtiene la relacion entre auto y dia por propiedades",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Propiedades de los objetos dia y auto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "licenseCar": {
                  "type": "string"
                },
                "dayTravel": {
                  "type": "integer"
                },
                "monthTravel": {
                  "type": "integer"
                },
                "yearTravel": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idRelationCarDate": {
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/relationcardate/getIdRelationCarDate": {
      "post": {
        "tags": [
          "Relacion Auto-Dia"
        ],
        "summary": "Obtiene la relacion entre auto y dia por ids",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ids de los objetos dia y auto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idCar": {
                  "type": "integer"
                },
                "idDay": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idRelationCarDate": {
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/relationcardate/getIdDatebyIdRelation/{id}": {
      "get": {
        "tags": [
          "Relacion Auto-Dia"
        ],
        "summary": "Busca el nodo dia por el id relacion",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de la relacion",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "idnodeDate": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  }
}